Factor out common utils
