+++ hypothesis-python/tests/conjecture/common.py
@@ -1,0 +1,87 @@
+# coding=utf-8
+#
+# This file is part of Hypothesis, which may be found at
+# https://github.com/HypothesisWorks/hypothesis/
+#
+# Most of this work is copyright (C) 2013-2019 David R. MacIver
+# (david@drmaciver.com), but it contains contributions by others. See
+# CONTRIBUTING.rst for a full list of people who may hold copyright, and
+# consult the git log if you need to determine who owns an individual
+# contribution.
+#
+# This Source Code Form is subject to the terms of the Mozilla Public License,
+# v. 2.0. If a copy of the MPL was not distributed with this file, You can
+# obtain one at https://mozilla.org/MPL/2.0/.
+#
+# END HEADER
+
+from __future__ import absolute_import, division, print_function
+
+from contextlib import contextmanager
+
+import hypothesis.internal.conjecture.engine as engine_module
+from hypothesis import HealthCheck, settings
+from hypothesis.internal.compat import hbytes
+from hypothesis.internal.conjecture.data import Status
+from hypothesis.internal.conjecture.engine import (
+    MIN_TEST_CALLS,
+    ConjectureRunner,
+    ExitReason,
+)
+from hypothesis.internal.conjecture.utils import (
+    Sampler,
+    calc_label_from_name,
+    integer_range,
+)
+from hypothesis.internal.entropy import deterministic_PRNG
+
+SOME_LABEL = calc_label_from_name("some label")
+
+
+TEST_SETTINGS = settings(
+    max_examples=5000, database=None, suppress_health_check=HealthCheck.all()
+)
+
+
+def run_to_data(f):
+    with deterministic_PRNG():
+        runner = ConjectureRunner(f, settings=TEST_SETTINGS)
+        runner.run()
+        assert runner.interesting_examples
+        (last_data,) = runner.interesting_examples.values()
+        return last_data
+
+
+def run_to_buffer(f):
+    return hbytes(run_to_data(f).buffer)
+
+
+@contextmanager
+def buffer_size_limit(n):
+    original = engine_module.BUFFER_SIZE
+    try:
+        engine_module.BUFFER_SIZE = n
+        yield
+    finally:
+        engine_module.BUFFER_SIZE = original
+
+
+def shrinking_from(start):
+    def accept(f):
+        with deterministic_PRNG():
+            runner = ConjectureRunner(
+                f,
+                settings=settings(
+                    max_examples=5000,
+                    database=None,
+                    suppress_health_check=HealthCheck.all(),
+                ),
+            )
+            runner.cached_test_function(start)
+            assert runner.interesting_examples
+            (last_data,) = runner.interesting_examples.values()
+            return runner.new_shrinker(
+                last_data, lambda d: d.status == Status.INTERESTING
+            )
+
+    return accept
