+++ hypothesis-python/tests/conjecture/test_conjecture_engine.py
@@ -18,7 +18,6 @@
 from __future__ import absolute_import, division, print_function
 
 import re
-from contextlib import contextmanager
 from random import Random
 
 import pytest
@@ -46,36 +45,14 @@ from hypothesis.internal.conjecture.utils import (
 from hypothesis.internal.entropy import deterministic_PRNG
 from tests.common.strategies import SLOW, HardToShrink
 from tests.common.utils import no_shrink
-
-SOME_LABEL = calc_label_from_name("some label")
-
-
-TEST_SETTINGS = settings(
-    max_examples=5000, database=None, suppress_health_check=HealthCheck.all()
+from tests.conjecture.common import (
+    SOME_LABEL,
+    TEST_SETTINGS,
+    buffer_size_limit,
+    run_to_buffer,
+    run_to_data,
+    shrinking_from,
 )
-
-
-def run_to_data(f):
-    with deterministic_PRNG():
-        runner = ConjectureRunner(f, settings=TEST_SETTINGS)
-        runner.run()
-        assert runner.interesting_examples
-        (last_data,) = runner.interesting_examples.values()
-        return last_data
-
-
-def run_to_buffer(f):
-    return hbytes(run_to_data(f).buffer)
-
-
-@contextmanager
-def buffer_size_limit(n):
-    original = engine_module.BUFFER_SIZE
-    try:
-        engine_module.BUFFER_SIZE = n
-        yield
-    finally:
-        engine_module.BUFFER_SIZE = original
 
 
 def test_can_index_results():
@@ -974,27 +951,6 @@ def test_try_shrinking_blocks_ignores_overrun_blocks(monkeypatch):
             data.mark_interesting()
 
     assert list(x) == [2, 2, 1]
-
-
-def shrinking_from(start):
-    def accept(f):
-        with deterministic_PRNG():
-            runner = ConjectureRunner(
-                f,
-                settings=settings(
-                    max_examples=5000,
-                    database=None,
-                    suppress_health_check=HealthCheck.all(),
-                ),
-            )
-            runner.cached_test_function(start)
-            assert runner.interesting_examples
-            (last_data,) = runner.interesting_examples.values()
-            return runner.new_shrinker(
-                last_data, lambda d: d.status == Status.INTERESTING
-            )
-
-    return accept
 
 
 def test_dependent_block_pairs_is_up_to_shrinking_integers():
